+++ js.bolt.surge/scripts/get-manifest.js
@@ -9,46 +9,18 @@ if (fs.realpathSync(process.argv[1]) === fileURLToPath(import.meta.url)) {
 }
 
 /**
- * Write source onto target to overwrite existing fields and create anew.
- */
-function merge(target, source) {
-  Object.keys(source).forEach((key) => {
-    if (key === "__proto__" || key === "constructor") return;
-    if (target.hasOwnProperty(key)) {
-      if (typeof target[key] === "object" && typeof source[key] === "object") {
-        merge(target[key], source[key]);
-      } else {
-        target[key] = source[key];
-      }
-    } else {
-      target[key] = source[key];
-    }
-  });
-  return target;
-}
-
-/**
- * Read and parse a manifest or throw in attempt.
- */
-function readManifest(filename) {
-  const pwd = process.env.PWD;
-  const manifest = path.join(pwd, `manifest/${filename}`);
-  if (fs.existsSync(manifest)) {
-    const file = fs.readFileSync(manifest, "utf8");
-    return JSON.parse(file);
-  }
-  throw new Error("No manifest file found");
-}
-
-/**
- * Retrieve a tagged manifest or fallback to a default.
+ * Retrieve a tagged manifest for the environment.
  */
 export default function getManifest() {
   const tag = process.env.SLACK_ENVIRONMENT_TAG;
   if (!tag) {
     throw new Error("No environment tag found");
   }
-  const manifest = readManifest(`manifest.json`);
-  const customizations = readManifest(`manifest.${tag}.json`);
-  return merge(manifest, customizations);
+  const pwd = process.env.PWD;
+  const manifest = path.join(pwd, `manifest.${tag}.json`);
+  if (fs.existsSync(manifest)) {
+    const file = fs.readFileSync(manifest, "utf8");
+    return JSON.parse(file);
+  }
+  throw new Error(`No manifest file found for environment: ${tag}`);
 }